Try changing a few types that represent sizes to size_t

+++ components/scream/src/physics/spa/atmosphere_prescribed_aerosol.hpp
@@ -82,7 +82,7 @@ protected:
   void finalize_impl   ();
 
   // Computes total number of bytes needed for local variables
-  int requested_buffer_size_in_bytes() const;
+  size_t requested_buffer_size_in_bytes() const;
 
   // Set local variables using memory provided by
   // the ATMBufferManager
